AI_Assist/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
PyYAML
python-json-logger
httpx
jinja2
python-dotenv